+++ classes/apixu_api.class.php
@@ -5,34 +5,21 @@
 *   @author     Lee Garner <lee@leegarner.com>
 *   @copyright  Copyright (c) 2016 Lee Garner <lee@leegarner.com>
 *   @package    weather
-*   @version    1.0.0
+*   @version    1.0.4
 *   @license    http://opensource.org/licenses/gpl-2.0.php 
 *               GNU Public License v2 or later
 *   @filesource
 */
 
+require_once dirname(__FILE__) . '/base_api.class.php';
+
 /**
-*   Class to manage Weather Underground
-*   @since  version 1.0.0
+*   Class to use apixu.com weather provider
+*   @since  version 1.0.r40
 *   @package weather
 */
-class Weather
+class Weather extends WeatherBase
 {
-    // Our variables are all available publically, though typically not used
-    public $response;
-    public $location;
-    public $current;
-    public $forecast;
-    public $fc_text;
-    public $info;
-    public $http_code;
-    public $error = 0;
-    public $api_name;
-    public $url;
-
-    private $have_fopen = false;
-    private $have_curl = false;
-
     /**
     *   Constructor.
     *   Get the weather for the location string, if specified
@@ -44,23 +31,14 @@ class Weather
         global $_CONF, $_CONF_WEATHER;
 
         $this->api_name = 'APIXU';
+        $this->api_code = 'apixu';
+        $this->max_days = 10;
 
-        if (empty($_CONF_WEATHER['api_key_apixu'])) {
-            if (SEC_inGroup('Root')) {
-                $this->error = WEATHER_ERR_KEYMISSING;
-            } else {
-                $this->error = WEATHER_ERR_API;
-            }
-            return;
-        }
-        $this->url = "https://api.apixu.com/v1/forecast.json?key={$_CONF_WEATHER['api_key_apixu']}&days=5&q=";
- 
-        if (in_array('curl', get_loaded_extensions())) {
-            // CURL is preferred since it handles other character sets better.
-            $this->have_curl = true;
-        } elseif (ini_get('allow_url_fopen') == 1) {
-            $this->have_fopen = true;
-        }
+        parent::__construct($loc);
+
+        $this->url = 'https://api.apixu.com/v1/forecast.json?key=' .
+                $_CONF_WEATHER['api_key_apixu'] .
+                 '&days=' . $this->fc_days . '&q=';
 
         if (!empty($loc)) {
             // Get the weather for the specified location, if requested
@@ -70,50 +48,19 @@ class Weather
 
 
     /**
-    *   Get the weather for a given location
-    *   If $loc is not specified, use the current saved location
+    *   Format a url for this provider.
+    *   Just appends a urlencoded location to the base url
     *
-    *   @param  string  $loc    Optional location to retrieve.
-    *   @return boolean     True on success, False on failure
+    *   @param  string  $loc    Location
+    *   @return string      Full API URL
     */
-    public function Get($loc = '')
+    protected function _makeUrl($loc)
     {
-        if (!empty($loc)) {
-            // Sanitize the location
-            $this->location = $loc;
-        }
-
-        if (empty($this->location)) {
-            COM_errorLog('Empty location provided');
-            $this->error = WEATHER_ERR_NOTFOUND;
-            return false;
-        }
-        $loc = rawurlencode($this->location);
-        $json = $this->GetWeather($this->url . $loc);
-        if (empty($json)) {
-            $this->error = WEATHER_ERR_API;
-            COM_errorLog('Empty weather data from ' . $this->api_name);
-            return false;
-        }
-
-        $A = json_decode($json);
-        if (!is_object($A)) {
-            $this->error = WEATHER_ERR_API;
-            COM_errorLog("error decoding json: $json");
-            return false;
-        }
-        $this->response = $A;
-
-        if (!$this->Parse()) {
-            COM_errorLog('error parsing json object');
-            $this->error = WEATHER_ERR_API;
-            return false;
-        }
-
-        return true;
+        $this->location = $loc;
+        return $this->url . rawurlencode($this->location);
     }
 
-
+         
     /**
     *   Parse the returned weather information.
     *   This function just puts the forecast info into some "shortcut"
@@ -146,63 +93,21 @@ class Weather
     
 
     /**
-    *   Fetch data from a remote server using php-curl
-    *
-    *   @param  string  $url    URL to retrieve
-    *   @return string      Data from website
-    */
-    private function GetWeather($url)
-    {
-        if ($this->have_curl) {
-            $agent = 'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-GB; ' .
-                'rv:1.9.1) Gecko/20090624 Firefox/3.5 (.NET CLR ' .
-                '3.5.30729)';
-
-            $ch = curl_init();
-            curl_setopt($ch, CURLOPT_URL,            $url);
-            curl_setopt($ch, CURLOPT_USERAGENT,      $agent);
-            curl_setopt($ch, CURLOPT_HEADER,         0);
-            curl_setopt($ch, CURLOPT_ENCODING,       'gzip');
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-            curl_setopt($ch, CURLOPT_FOLLOWLOCATION, 1);
-            curl_setopt($ch, CURLOPT_FAILONERROR,    1);
-            curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, 8);
-            curl_setopt($ch, CURLOPT_TIMEOUT,        8);
-            /*curl_setopt($ch, CURLOPT_HTTPHEADER, array(
-                    'Accept-Charset: utf-8',
-                ) );*/
-            //curl_setopt($ch, CURLOPT_VERBOSE,        1);
-
-            $result = curl_exec($ch);
-            $this->http_code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-            curl_close($ch);
-            if ($this->http_code != '200') {
-                $result = '';
-            }
-
-        } elseif ($this->have_fopen) {
-            $result = file_get_contents($url, 0);
-        } else {
-            $result = '';
-            COM_errorLog('WEATHER: Missing url_fopen and curl support');
-        }
-        return $result;
-    }
-
-
-    /**
     *   Get the data into standard arrays for cache storage and display.
     *   Collects data from the info, current and forecast variables
-    *   depending on the format of the weather provider
+    *   depending on the format of the weather provider.
+    *
+    *   Returns the data array as well as setting the values in $this->data
+    *   to be publicly consumed.
     *
     *   @return array   Data array
     */
     public function getData()
     {
-        global $_USER;
+        global $_USER, $_CONF_WEATHER;
 
         //list($city, $country) = explode(', ', $this->info->query);
-        $data = array(
+        $this->data = array(
             'info' => array(
                 'city'  => $this->info,
                 'date_time' => $this->response->location->localtime,
@@ -228,15 +133,14 @@ class Weather
             if (empty($this->response->location->tz_id)) {
                 $this->response->location->tz_id = $_USER['tzid'];
             }
-            //$D = new Date('now', $this->response->location->tz_id);
-            for ($i = 0; $i < 5; $i++) {
-                // We want 5 days.
+            // Create Date object for getting the day name from the timestamp
+            $D = new Date('now', $this->response->location->tz_id);
+            for ($i = 0; $i < $this->fc_days; $i++) {
                 if (!isset($this->forecast[$i])) break;
                 $fc = $this->forecast[$i];
-                //$t = 2 * $i;    // index into text descriptions, 2 per entry
-              //  $D->setTimestamp($fc->date_epoch);
-                $data['forecast'][] = array(
-                //    'day'    => $D->format('l', true),
+                $D->setTimestamp($fc->date_epoch);
+                $this->data['forecast'][] = array(
+                    'day'    => $D->format('l', true),
                     'lowF'   => (string)$fc->day->mintemp_f,
                     'highF'  => (string)$fc->day->maxtemp_f,
                     'lowC'   => (string)$fc->day->mintemp_c,
@@ -251,36 +155,7 @@ class Weather
                 );
             }
         }
-        $this->data = $data;
-        return $data;
-    }
-
-
-    /**
-    *   Return the linkback url to the weather provider.
-    *   Weather Underground requires this for the free API
-    *
-    *   @param  string  $format     'page' or 'block' for horiz or vert image
-    *   @return string  Linkback tag
-    */
-    public static function linkback($format='page')
-    {
-        return '';
-    }
-
-
-    /**
-    *   Return the URL for a give icon string.
-    *
-    *   @param  string  $icon   Icon URL returned from the weather API
-    *   @return string          Fully-qualified URL to the icon image
-    */
-    public static function getIcon($icon)
-    {
-        if (function_exists('CUSTOM_weatherIcon')) {
-            $icon = CUSTOM_weatherIcon($icon);
-        }
-        return $icon;
+        return $this->data;
     }
 
 }   // class Weather