+++ language/german_formal_utf-8.php
@@ -40,7 +40,7 @@ $LANG_fs['weather'] = array(
     'fs_main'               => 'Allgemeine Einstellungen',
     'fs_provider_wwo'       => 'World Weather Online',
     'fs_provider_wu'        => 'Weather Underground',
-    'fs_provider_axpiu'     => 'AXPIU',
+    'fs_provider_apixu'     => 'APIXU',
 );
 
 $LANG_configsections['weather'] = array(
@@ -62,6 +62,7 @@ $LANG_confignames['weather'] = array(
     'api_key_wwo'       => 'API-Key',
     'api_key_wu'        => 'API-Key',
     'ref_key_wu'        => 'Referrer-Key',
+    'api_key_apixu'     => 'API-Key',
 );
 
 $LANG_configselects['weather'] = array(
@@ -84,6 +85,7 @@ $LANG_configselects['weather'] = array(
     15  => array(   'Farenheit' => 'F', 'Celsius' => 'C'),
     16  => array(   'World Weather Online' => 'wwo',
                     'Weather Underground' => 'wu',
+                    'APIXU' => 'apixu',
             ),
 );