+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 1.0.4 -
 - Add apixu.com weather provider
+- Add changes from matrox66 for World Weather Online class
+- Extract common weather functions into base_api.class.php
 
 ## Version 1.0.3 - 2016-11-06
 - Hide large description from display on small viewports